Memisahkan class Child dari Employee untuk mengurangi kompleksitas class
